Moved ImGui renderer to a separate file so we can more easily reuse it later.
